Update README.md
PRINTF PROJECT

Description
In C programming language, printf() is a standard library function that is used to print or display formatted output to the console or standard output stream (stdout). It stands for "print formatted" and is declared in the header file stdio.h.
The printf() function takes a format string as its first argument, followed by a variable number of arguments that correspond to the placeholders in the format string. The format string contains both literal text and format specifiers, which are used to control the display of the values passed in as arguments.
This time, in order to comprehend the complexity of a potent tool, we are showing a printf project that was created from scratch in the majority of its features. To learn how _printf functions, go to man 3 of printf.

Compilation
$ gcc -Wall -Werror -Wextra -pedantic *.c
Prototype
int _printf(const char *format, ...)
Return
The function returns the total amount of characters produced if everything goes according to plan.
Formats
Specifiers	Description
%c	Prints single character
%s	Prints string of characters
%d	Prints decimal numbers
%i	Prints integers